+++ source/GUI/Components/Rectangle.cpp
@@ -1,0 +1,13 @@
+#include "Rectangle.hpp"
+
+
+Rect::Rect(vec2 position, vec2 dimensions, colRGB color)
+	: pos(position)
+	, dim(dimensions)
+	, col(color)
+{
+}
+Rect::Rect()
+{
+
+}