+++ source/GUI/Panel.cpp
@@ -1,0 +1,11 @@
+#include "Panel.hpp"
+
+
+Panel::Panel()
+{
+}
+
+
+Panel::~Panel()
+{
+}